requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.1
Django==1.11.11
django-dotenv==1.4.1
facebook-sdk==2.0.0
gunicorn==19.6.0
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
whitenoise==3.2
